# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: User Authentication API
  description: |
    Production-ready authentication API with JWT tokens, MongoDB integration, email verification, and password reset functionality.

    ## Features
    - User registration with email verification
    - JWT-based authentication (access + refresh tokens)
    - Password reset via email
    - Secure password hashing with bcrypt
    - Rate limiting and security middleware
    - Comprehensive input validation

    ## Security
    - All endpoints use HTTPS in production
    - JWT tokens for stateless authentication
    - Password hashing with bcrypt (12 rounds)
    - Input validation and sanitization
    - Rate limiting on sensitive endpoints
    - XSS and NoSQL injection protection

    ## Authentication
    Most endpoints require a valid JWT access token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.example.com/api/v1
    description: Production server
...
paths:
...
  /auth/verify-email:
    post:
      tags:
        - Authentication
      summary: Verify email address
      description: Verify user email using token sent via email
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  example: abc123def456ghi789jkl012mno345pqr678stu901vwx234yz
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Email verified successfully. You can now login.
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/resend-verification:
    post:
      tags:
        - Authentication
      summary: Resend verification email
      description: Send a new verification email to the user
      operationId: resendVerification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: john.doe@example.com
      responses:
        '200':
          description: Verification email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Verification email sent. Please check your inbox.
        '400':
          description: Email already verified or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...